requirements.txt
pylibdmtx>=0.1.10
numpy>=1.21.0
Pillow>=8.0.0
PyQt5>=5.15.0
